Move all subcommands to their own files and fns. This is only a move.  No logic should have changed apart from the addition of VERBOSE and REPEAT globar vars and from some clarified restructuring of params argument interpretation.
